requirements.txt
ccxt==4.4.26
vectorbt>=0.26
ta>=0.11.0
pandas>=2.1.0,<3.0
numpy>=1.24.0
apscheduler==3.10.4
sqlalchemy==2.0.36
anthropic>=0.40.0
python-telegram-bot==21.7
pydantic-settings==2.6.1
pyyaml>=6.0
sentence-transformers==3.3.1
memvid-sdk>=2.0
requests>=2.31.0
python-dotenv>=1.0.0

# Dev dependencies
pytest>=7.0
pytest-asyncio
